- Fix: create 시 중복 체크 로직 추가 - Fix: Message 업데이트 시 부모 message와 isReply 체크

+++ src/main/java/com/sprint/mission/discodeit/service/file/FileChanelCrudService.java
@@ -30,6 +30,10 @@ public class FileChanelCrudService implements ChannelService {
     @Override
     public void create(Channel channel) {
 
+        if (existById(channel.getId())) {
+            throw new IllegalArgumentException("Channel already exists.");
+        }
+
         try(FileOutputStream fos = new FileOutputStream(path.resolve( channel.getId() + FILE_EXTENSION).toFile());
             ObjectOutputStream oos = new ObjectOutputStream(fos)) {
